fix(edge/scanner): cast csdOK to C.LONG for CGO type compatibility
CGO does not implicitly convert typed Go int32 constants to C.LONG.
Explicit C.LONG() cast on the CsdParSet(CSDP_DBLFEEDSTATUS) call.

CGO 32-bit build now clean: GOARCH=386 + i686-w64-mingw32-gcc 16.1.0
Binary verified: 32-bit i386 PE (~13.5 MB), 17/17 tests passing.

+++ edge/cts-scanner-agent/ranger_windows.go
@@ -476,7 +476,7 @@ func (t *CanonTransport) ReadItem() (*ScannedItem, error) {
 				C.astra_release_image(&frontImg)
 				hasFront = false
 			}
-			C.astra_par_set_long(C.CSDP_DBLFEEDSTATUS, csdOK)
+			C.astra_par_set_long(C.CSDP_DBLFEEDSTATUS, C.LONG(csdOK))
 			t.logger.Warn("double-feed detected by ultrasonic sensor")
 			return &ScannedItem{DoubleFeedDetected: true}, nil
